+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
   * Added contextual menu
   * Bottom bar on hover style fixed
+  * Improved tabs responsiveness
   * Fixed bug where more than the max tasks per suite could be added or dragged
   * Increased timer event emitter max listeners to remove warning
   * Fixed bug with infinite tasks printing to stdout