+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Fixed bug where more than the max tasks per suite could be added or dragged
   * Increased timer event emitter max listeners to remove warning
   * Fixed bug with infinite tasks printing to stdout
+  * Minor styles improvements in tabs
 
 0.5.0 / 2018-06-03
 ==================